List an Associative array
This was an easy task,
the problem where I´m stuck is,
I want to update the array, for example the quantity of the Bananas by using the user input,
I think I can do it with readline() using the console. (it's more similar do C)
But since I'm learning I want to have a form with buttons, so I need to learn HTML and Javascript I think...
and maybe have a SQL server.. Still trying to understand what I actually need.

Feeling dumb xD